refactor(SetProfile): Move profile picking to util for consistency
