+++ src/nr1-cli-commands.ts
@@ -1,6 +1,3 @@
-const fs = require("fs");
-const os = require("os");
-
 import * as vscode from "vscode";
 
 import runCommand from "./utils/run-command";
@@ -25,27 +22,5 @@ export const catalogInfo = () => "nr1 catalog:info";
 
 export const catalogSubmit = () => "nr1 catalog:submit";
 
-const setProfile = (profileName: string | undefined) =>
+export const setProfile = (profileName: string) =>
   `nr1 profiles:default -n ${profileName?.replace(" (current)", "")}`;
-
-export const selectProfile = async () => {
-  const credentialPath = `${os.homedir()}/.newrelic/credentials.json`;
-  const defaultPath = `${os.homedir()}/.newrelic/default-profile.json`;
-
-  const profiles = JSON.parse(fs.readFileSync(credentialPath));
-  const currentDefault = JSON.parse(fs.readFileSync(defaultPath));
-  const profileNames = Object.keys(profiles).map((profileName) => {
-    if (profileName === currentDefault) {
-      return `${profileName} (current default)`;
-    }
-    return profileName;
-  });
-  const profileName = await vscode.window.showQuickPick(profileNames);
-
-  runCommand(setProfile(profileName), () => {
-    profileName &&
-      vscode.window.showInformationMessage(
-        `Default profile updated to ${profileName}`
-      );
-  });
-};
